Rename to containerWith
