+++ src/expect.ts
@@ -1,7 +1,7 @@
 import { Expectation, ExpectationOptions } from './Expectation'
 import { AMatcher } from './matchers/A'
 import { AnythingMatcher } from './matchers/Anything'
-import { ArrayContainingMatcher } from './matchers/ArrayContaining'
+import { ContainerWithMatcher } from './matchers/ContainerWith'
 import { NumberCloseToMatcher } from './matchers/NumberCloseTo'
 import { StringMatchingMatcher } from './matchers/StringMatching'
 
@@ -13,7 +13,7 @@ interface expectInterface {
   a: typeof AMatcher.make
   stringMatching: typeof StringMatchingMatcher.make
   numberCloseTo: typeof NumberCloseToMatcher.make
-  arrayContaining: typeof ArrayContainingMatcher.make
+  containerWith: typeof ContainerWithMatcher.make
 }
 
 export const expect: expectInterface = <T>(actual: T, options: ExpectationOptions = {}): Expectation<T> => {
@@ -23,4 +23,4 @@ expect.anything = AnythingMatcher.make
 expect.a = AMatcher.make
 expect.stringMatching = StringMatchingMatcher.make
 expect.numberCloseTo = NumberCloseToMatcher.make
-expect.arrayContaining = ArrayContainingMatcher.make
+expect.containerWith = ContainerWithMatcher.make
